LayoutTests/fast/dom/Document/scrollingElement-quirks-mode-expected.txt

Tests the behavior of document.scrollingElement in quirks mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


In quirks mode the scrolling element is `BODY`
PASS document.scrollingElement is document.body
PASS document.scrollingElement is document.body
PASS document.scrollingElement is document.body
In quirks mode, if the `BODY` is scrollable, the scrolling element is `null`
PASS document.scrollingElement is null
In quirks mode, if the `BODY` is `display: none`, the scrolling element is `BODY`
PASS document.scrollingElement is document.body
In quirks mode, the document.scrollingElement is null if there is no `HTML` root.
PASS scrollingElement is null
PASS successfullyParsed is true

TEST COMPLETE
